\name{plot_raster}
\alias{plot_raster}
\title{Simple raster plotting for sleep data.}
\description{
  This function creates a simple raster plot for sleep data given in a JSON file.
}
\usage{
  plot_raster(json_path)
}
\arguments{
  \item{json_path}{is the path of the json input file that contains data to be plotted.}
}
